WIP: quick and dirty refactor to decision service
